Random: move irand and drand to common
