+++ src/common/CMakeLists.txt
@@ -24,7 +24,7 @@ target_sources(common
 install(FILES
         Integer.h Number.h FastRational.h XAlloc.h Alloc.h StringMap.h Timer.h osmtinttypes.h
         TreeOps.h Real.h FlaPartitionMap.h PartitionInfo.h OsmtApiException.h TypeUtils.h 
-        NumberUtils.h NatSet.h
+        NumberUtils.h NatSet.h Random.h
 DESTINATION ${INSTALL_HEADERS_DIR})
